requirements.txt
click
flake8
biopython
nose
pysam
pyvcf
xmltodict
numpy
